+++ GaloisFieldHillCipher/include/EncryptionMatrixGeneration.h
@@ -5,4 +5,4 @@
 
 #include "StatusCodes.h"
 
-STATUS_CODE generate_encryption_matrix(long double*** out_matrix, uint32_t dimentation, uint32_t prime_field);
+STATUS_CODE generate_encryption_matrix(double*** out_matrix, uint32_t dimentation, uint32_t prime_field);
